% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observations}
\alias{observations}
\title{Stem height observations from 2017 Lythrum salicaria common garden}
\format{
A data frame with 21152 rows and 3 variables
}
\usage{
data(observations)
}
\description{
Stem height observations from 2017 Lythrum salicaria common garden
}
\keyword{datasets}
