% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Parameter estimates for individuals in common garden}
\format{
A data frame with 2278 rows and 11 variables
}
\usage{
data(parameters)
}
\description{
Parameter estimates for individuals in common garden
}
\keyword{datasets}
