% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predicted}
\alias{predicted}
\title{Predicted values from nonlinear mixed-effects model}
\format{
A data frame with 182240 rows and 6 variables
}
\usage{
data(predicted)
}
\description{
Predicted values from nonlinear mixed-effects model
}
\keyword{datasets}
